fix(guard): extract manifest dependency targets module
